refactor shit to poetry
