+++ .dumbjump
@@ -1,0 +1,6 @@
++/home/vlad/.cache/pypoetry/virtualenvs/house-prices-mlp-L7SDcTsr-py3.8/lib/python3.8/site-packages
++/home/vlad/.cache/pypoetry/virtualenvs/house-prices-mlp-L7SDcTsr-py3.8/local/lib/python3.8/dist-packages
++/home/vlad/.cache/pypoetry/virtualenvs/house-prices-mlp-L7SDcTsr-py3.8/lib/python3/dist-packages
++/home/vlad/.cache/pypoetry/virtualenvs/house-prices-mlp-L7SDcTsr-py3.8/lib/python3.8/dist-packages
+-tests
+-*.
