+++ .gitignore
@@ -5,3 +5,4 @@ data/
 logs/
 typings/
 *.csv
+*~
