+++ pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "house_prices_mlp"
+version = "0.1.0"
+description = ""
+authors = ["Arnaud Autef <arnaud.autef@gmail.com>", "Vladimir Feinberg <vladimir.feinberg@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+tensorflow-cpu = "^2.6.0"
+jax = {extras = ["cpu"], version = "^0.2.21"}
+jupyter = "^1.0.0"
+flax = "^0.3.5"
+pandas = "^1.3.3"
+
+[tool.poetry.dev-dependencies]
+pytest = "^6.2.5"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
